# clinic-api/docs/swagger.yaml
basePath: /api
definitions:
...
  payment.CreateInput:
    properties:
      amount:
        description: Amount is in cents (e.g. 15000 = R$ 150,00). Must be greater
          than zero.
        example: 15000
        type: integer
      clinic_id:
        example: a0000000-0000-0000-0000-000000000001
        type: string
      dentist_id:
        type: string
    type: object
  payment.listResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/payment.paymentResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  payment.paymentResponse:
    properties:
      amount:
        example: 15000
        type: integer
      approved_at:
        type: string
      clinic_id:
        example: a0000000-0000-0000-0000-000000000001
        type: string
      created_at:
        type: string
      dentist_id:
        type: string
      id:
        example: 62818880-fffa-4512-bb48-34692f49e3cb
        type: string
      pix_code:
        example: 000201262yZEd13q_BCi_wyx1gxaDLUbQfdU7oY5dm7GMumLL_A
        type: string
      status:
        enum:
        - pending
        - approved
        example: pending
        type: string
      updated_at:
        type: string
    type: object
  problem.Details:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      detail:
        example: request validation failed
        type: string
      errors:
        items:
          $ref: '#/definitions/problem.FieldError'
        type: array
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
      type:
        example: about:blank
        type: string
    type: object
  problem.FieldError:
    properties:
      detail:
        example: is required
        type: string
      field:
        example: clinic_id
        type: string
    type: object
info:
  contact: {}
  description: |-
    API de gestão de clínica odontológica: clínicas, dentistas e pagamentos via Pix (simulado).
    Todos os endpoints de negócio ficam sob o prefixo /api; o único fora dele é o health check em /health.
    Erros seguem o formato RFC 9457 (application/problem+json) em todo o corpo, com um campo "code" estável para tratamento programático além do "status" HTTP.
  title: Clinic API
  version: "1.0"
paths:
...
  /payments:
    get:
      parameters:
      - description: Clinic ID
        format: uuid
        in: query
        name: clinic_id
        required: true
        type: string
      - description: Filter by status
        enum:
        - pending
        - approved
        in: query
        name: status
        type: string
      - default: 20
        description: Page size
        in: query
        maximum: 100
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.listResponse'
        "400":
          description: clinic_id missing or not a valid UUID
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: clinic not found
          schema:
            $ref: '#/definitions/problem.Details'
      summary: List payments for a clinic (paginated)
      tags:
      - payments
    post:
      consumes:
      - application/json
      description: Creates a "pending" payment with a simulated Pix copy-and-paste
        code. A background job approves it automatically after a random 2-5s delay.
        Replaying the same Idempotency-Key with the same body returns the original
        payment (200); replaying it with a different body returns 409.
      parameters:
      - description: Client-generated unique key; safe to retry a request with the
          same key
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Payment to create
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/payment.CreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: replay of an existing payment (same Idempotency-Key and body)
          schema:
            $ref: '#/definitions/payment.paymentResponse'
        "201":
          description: payment created
          schema:
            $ref: '#/definitions/payment.paymentResponse'
        "400":
          description: missing Idempotency-Key header or validation error
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: clinic or dentist not found
          schema:
            $ref: '#/definitions/problem.Details'
        "409":
          description: clinic not active, or Idempotency-Key reused with a different
            body
          schema:
            $ref: '#/definitions/problem.Details'
      summary: Create a Pix payment
      tags:
      - payments
  /payments/{id}:
    get:
      description: Poll this endpoint to observe the status transition from "pending"
        to "approved".
      parameters:
      - description: Payment ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.paymentResponse'
        "400":
          description: id is not a valid UUID
          schema:
            $ref: '#/definitions/problem.Details'
        "404":
          description: payment not found
          schema:
            $ref: '#/definitions/problem.Details'
      summary: Get a payment by id
      tags:
      - payments
schemes:
- http
swagger: "2.0"