# clinic-api/internal/payment/handler.go
package payment

import (
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"
	"strconv"
	"strings"

	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"

	"github.com/josuesantos1/desafio/pkg/problem"
)

func RegisterRoutes(r chi.Router, svc *Service) {
	h := &handler{svc: svc}
	r.Post("/payments", h.create)
	r.Get("/payments", h.list)
	r.Get("/payments/{id}", h.get)
}

type handler struct {
	svc *Service
}

// @Summary      Create a Pix payment
// @Description  Creates a "pending" payment with a simulated Pix copy-and-paste code. A background job approves it automatically after a random 2-5s delay. Replaying the same Idempotency-Key with the same body returns the original payment (200); replaying it with a different body returns 409.
// @Tags         payments
// @Accept       json
// @Produce      json
// @Param        Idempotency-Key  header    string       true  "Client-generated unique key; safe to retry a request with the same key"
// @Param        payment          body      CreateInput  true  "Payment to create"
// @Success      201              {object}  paymentResponse  "payment created"
// @Success      200              {object}  paymentResponse  "replay of an existing payment (same Idempotency-Key and body)"
// @Failure      400              {object}  problem.Details  "missing Idempotency-Key header or validation error"
// @Failure      404              {object}  problem.Details  "clinic or dentist not found"
// @Failure      409              {object}  problem.Details  "clinic not active, or Idempotency-Key reused with a different body"
// @Router       /payments [post]
func (h *handler) create(w http.ResponseWriter, r *http.Request) {
	key := r.Header.Get("Idempotency-Key")
	if strings.TrimSpace(key) == "" {
		problem.Write(w, http.StatusBadRequest, "IDEMPOTENCY_KEY_REQUIRED", "Idempotency-Key header is required", nil)
		return
	}

	var in CreateInput
	if err := json.NewDecoder(r.Body).Decode(&in); err != nil {
		problem.Write(w, http.StatusBadRequest, "VALIDATION_ERROR", "invalid request body", nil)
		return
	}

	p, created, err := h.svc.Create(r.Context(), key, in)
	if err != nil {
		writeDomainError(w, err)
		return
	}

	status := http.StatusOK
	if created {
		status = http.StatusCreated
	}
	problem.WriteJSON(w, status, toPaymentResponse(p))
}

// @Summary      Get a payment by id
// @Description  Poll this endpoint to observe the status transition from "pending" to "approved".
// @Tags         payments
// @Produce      json
// @Param        id   path      string  true  "Payment ID"  format(uuid)
// @Success      200  {object}  paymentResponse
// @Failure      400  {object}  problem.Details  "id is not a valid UUID"
// @Failure      404  {object}  problem.Details  "payment not found"
// @Router       /payments/{id} [get]
func (h *handler) get(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	if _, err := uuid.Parse(id); err != nil {
		problem.Write(w, http.StatusBadRequest, "INVALID_ID", "id must be a valid UUID", nil)
		return
	}

	p, err := h.svc.Get(r.Context(), id)
	if err != nil {
		writeDomainError(w, err)
		return
	}

	problem.WriteJSON(w, http.StatusOK, toPaymentResponse(p))
}

// @Summary      List payments for a clinic (paginated)
// @Tags         payments
// @Produce      json
// @Param        clinic_id  query     string  true   "Clinic ID"  format(uuid)
// @Param        status     query     string  false  "Filter by status"  Enums(pending, approved)
// @Param        limit      query     int     false  "Page size"  default(20)  maximum(100)
// @Param        offset     query     int     false  "Offset for pagination"  default(0)
// @Success      200        {object}  listResponse
// @Failure      400        {object}  problem.Details  "clinic_id missing or not a valid UUID"
// @Failure      404        {object}  problem.Details  "clinic not found"
// @Router       /payments [get]
func (h *handler) list(w http.ResponseWriter, r *http.Request) {
	clinicID := r.URL.Query().Get("clinic_id")
	if clinicID == "" {
		problem.Write(w, http.StatusBadRequest, "VALIDATION_ERROR", "request validation failed",
			map[string]string{"clinic_id": "is required"})
		return
	}
	if _, err := uuid.Parse(clinicID); err != nil {
		problem.Write(w, http.StatusBadRequest, "INVALID_ID", "clinic_id must be a valid UUID", nil)
		return
	}

	params := ListParams{
		Limit:    clampLimit(parseQueryInt(r, "limit")),
		Offset:   clampOffset(parseQueryInt(r, "offset")),
		ClinicID: clinicID,
		Status:   parseQueryStatusPtr(r),
	}

	result, err := h.svc.List(r.Context(), params)
	if err != nil {
		writeDomainError(w, err)
		return
	}

	problem.WriteJSON(w, http.StatusOK, toListResponse(result, params))
}

func parseQueryInt(r *http.Request, key string) int {
	v, err := strconv.Atoi(r.URL.Query().Get(key))
	if err != nil {
		return 0
	}
	return v
}

func parseQueryStatusPtr(r *http.Request) *string {
	raw := r.URL.Query().Get("status")
	if raw != StatusPending && raw != StatusApproved {
		return nil
	}
	return &raw
}

func writeDomainError(w http.ResponseWriter, err error) {
	var validationErr *ValidationError
	switch {
	case errors.As(err, &validationErr):
		problem.Write(w, http.StatusBadRequest, "VALIDATION_ERROR", "request validation failed", validationErr.Fields)
	case errors.Is(err, ErrClinicNotFound):
		problem.Write(w, http.StatusNotFound, "CLINIC_NOT_FOUND", "clinic not found", nil)
	case errors.Is(err, ErrClinicNotActive):
		problem.Write(w, http.StatusConflict, "CLINIC_NOT_ACTIVE", "clinic must be active (have an administrator and a legal representative) to receive payments", nil)
	case errors.Is(err, ErrDentistNotFound):
		problem.Write(w, http.StatusNotFound, "DENTIST_NOT_FOUND", "dentist not found", nil)
	case errors.Is(err, ErrIdempotencyKeyConflict):
		problem.Write(w, http.StatusConflict, "IDEMPOTENCY_KEY_CONFLICT", "idempotency key already used with a different payload", nil)
	case errors.Is(err, ErrNotFound):
		problem.Write(w, http.StatusNotFound, "PAYMENT_NOT_FOUND", "payment not found", nil)
	default:
		slog.Error("unclassified payment repository error", "error", err)
		problem.Write(w, http.StatusInternalServerError, "INTERNAL_ERROR", "internal server error", nil)
	}
}